requirements.txt
asgiref==3.3.0
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
